[M487] Fix some compile warnings

+++ targets/TARGET_NUVOTON/TARGET_M480/device/TOOLCHAIN_GCC_ARM/m480_retarget.c
@@ -29,7 +29,7 @@ void *__wrap__sbrk(int incr)
     uint32_t heap_ind_old = NU_ALIGN_UP(heap_ind, NU_HEAP_ALIGN);
     uint32_t heap_ind_new = NU_ALIGN_UP(heap_ind_old + incr, NU_HEAP_ALIGN);
     
-    if (heap_ind_new > &__mbed_krbs_start) {
+    if (heap_ind_new > (uint32_t) &__mbed_krbs_start) {
         errno = ENOMEM;
         return (void *) -1;
     }
